+++ native/xqlitenif/src/lib.rs
@@ -1,4 +1,4 @@
-#![warn(clippy::undocumented_unsafe_blocks)]
+#![deny(clippy::undocumented_unsafe_blocks)]
 
 pub(crate) mod atoms {
     rustler::atoms! {